Make JKASSERT fail less likely in bot_combat.cpp:add_next_posdata()
Only possibility for this assert to fail is that all posdata_mem entries
have current or future time in time-field. Future time here is not possible,
it must be left-over over map-change. Issue ugly fix to replace future time
with current-time. It will cause short period of problems for bot's aiming,
but that should last only under second and happen once per map-change when
this bug is triggered.

Signed-off-by: Jussi Kivilinna <jussi.kivilinna@mbnet.fi>